requirements.txt
fastapi
uvicorn
requests
numpy
scikit-learn==1.3.2
tensorflow-cpu
joblib
python-multipart
# Add these two new lines
opencv-python-headless
librosa
